pyproject.toml
[project]
name = "franka-control-client"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "pytest>=8.3.5",
    "pyzmq>=26.4.0",
]
